Fix delete method for GoIds
Check that there is a file name before deleting in case the
user doesn't upload a sections file

+++ goatoolsgui/models.py
@@ -18,8 +18,10 @@ class GoIds(models.Model):
 
   def delete(self):
     # Model TODO: Going to need some checks here
-    print(self.sections_file.name)
-    os.remove(self.sections_file.name)
+    # print(self.sections_file.name)
+    # TODO: Handle case where no sections file exists
+    if self.sections_file.name != '':
+      os.remove(self.sections_file.name)
     print(self.file_out_name)
     os.remove(self.file_out_name)
     super(GoIds, self).delete()
